refactor(seed): split CI deployments out of ci.py, fix a vacuous window test
Moves the deterministic weighted-pick helper to base.py (it carries no CI
knowledge) and CI_WINDOW_DAYS/the retention clamp to ci_topology.py (which
already owns vendor CI facts), so ci.py holds one noun again and
ci_deployments.py holds the other, with no import cycle between them.

The window-clamp shape test previously seeded exactly CI_WINDOW_DAYS worth
of history, so the clamp had nothing to reject and would pass unchanged
even if removed. Widens the fixture history past the window and adds a
test that guards the widening itself, then confirms the clamp test now
fails when the clamp is neutralized.
